Issue #SB-26129 bug fix: App icon is stored as google drive url link on reading content in dock.
