+++ content-api/content-service/conf/application.conf
@@ -615,8 +615,6 @@ import {
   request_size_limit=2
   output_topic_name="local.auto.creation.job.request"
   required_props=["name","code","mimeType","contentType","artifactUrl","framework", "channel"]
-  g_service_acct_cred =
-  gdrive.application_name=drive-download
   initial_backoff_delay=120000
   maximum_backoff_delay=1200000
   increment_backoff_delay=2
