+++ platform-modules/url-manager/src/main/java/org/sunbird/url/util/GoogleDriveUrlUtil.java
@@ -1,6 +1,7 @@
 package org.sunbird.url.util;
 
 import com.google.api.client.googleapis.json.GoogleJsonResponseException;
+import com.google.api.client.http.HttpResponseException;
 import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.client.json.JsonFactory;
@@ -11,11 +12,14 @@ import com.google.api.services.drive.model.File;
 import org.apache.commons.collections4.MapUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.sunbird.common.Platform;
+import org.sunbird.common.Slug;
 import org.sunbird.common.exception.ClientException;
 import org.sunbird.common.exception.ServerException;
 import org.sunbird.telemetry.logger.TelemetryManager;
 import org.sunbird.url.common.URLErrorCodes;
 
+import java.io.FileOutputStream;
+import java.io.OutputStream;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -42,6 +46,10 @@ public class GoogleDriveUrlUtil {
 	private static final List<String> ERROR_CODES = Arrays.asList("dailyLimitExceeded402", "limitExceeded",
 			"dailyLimitExceeded", "quotaExceeded", "userRateLimitExceeded", "quotaExceeded402", "keyExpired",
 			"keyInvalid");
+	public static final Integer INITIAL_BACKOFF_DELAY = Platform.config.hasPath("import.initial_backoff_delay") ? Platform.config.getInt("import.initial_backoff_delay") : 1200000;    // 20 min
+	public static final Integer MAXIMUM_BACKOFF_DELAY = Platform.config.hasPath("import.maximum_backoff_delay") ? Platform.config.getInt("import.maximum_backoff_delay") : 3900000;    // 65 min
+	public static final Integer INCREMENT_BACKOFF_DELAY = Platform.config.hasPath("import.increment_backoff_delay") ? Platform.config.getInt("import.increment_backoff_delay") : 300000; // 5 min
+	public static Integer BACKOFF_DELAY = INITIAL_BACKOFF_DELAY;
 
 	private static boolean limitExceeded = false;
 	private static Drive drive = null;
@@ -137,5 +145,59 @@ public class GoogleDriveUrlUtil {
 		if (limitExceeded)
 			throw new ServerException(URLErrorCodes.ERR_GOOGLE_SERVICE.name(), SERVICE_ERROR);
 		return googleDriveFile;
+	}
+
+
+	public static java.io.File downloadFile(String fileId, String saveDir) {
+		try {
+			Drive.Files.Get getFile = drive.files().get(fileId);
+			getFile.setFields("id,name,size,owners,mimeType,properties,permissionIds,webContentLink");
+			com.google.api.services.drive.model.File googleDriveFile = getFile.execute();
+			TelemetryManager.log("GoogleDriveUtil :: downloadFile ::: Drive File Details:: " + googleDriveFile);
+			String fileName = googleDriveFile.getName();
+			java.io.File saveFile = new java.io.File(saveDir);
+			if (!saveFile.exists()) {
+				saveFile.mkdirs();
+			}
+			String saveFilePath = saveDir + java.io.File.separator + fileName;
+			TelemetryManager.log("GoogleDriveUtil :: downloadFile :: File Id :" + fileId + " | Save File Path: " + saveFilePath);
+			OutputStream outputStream = new FileOutputStream(saveFilePath);
+			getFile.executeMediaAndDownloadTo(outputStream);
+			outputStream.close();
+			java.io.File file = new java.io.File(saveFilePath);
+			file = Slug.createSlugFile(file);
+			TelemetryManager.log("GoogleDriveUtil :: downloadFile :: File Downloaded Successfully. Sluggified File Name: " + file.getAbsolutePath());
+			if (null != file && BACKOFF_DELAY != INITIAL_BACKOFF_DELAY)
+				BACKOFF_DELAY = INITIAL_BACKOFF_DELAY;
+			return file;
+		} catch(GoogleJsonResponseException ge) {
+			TelemetryManager.log("GoogleDriveUtil :: downloadFile :: GoogleJsonResponseException :: Error Occurred while downloading file having id "+fileId + " | Error is ::"+ge.getDetails().toString());
+			throw new ServerException(URLErrorCodes.ERR_INVALID_UPLOAD_FILE_URL.name(), "Invalid Response Received From Google API for file Id : " + fileId + " | Error is : " + ge.getDetails().toString());
+		} catch(HttpResponseException he) {
+			TelemetryManager.log("GoogleDriveUtil :: downloadFile :: HttpResponseException :: Error Occurred while downloading file having id "+fileId + " | Error is ::"+he.getContent());
+			he.printStackTrace();
+			if(he.getStatusCode() == 403) {
+				if (BACKOFF_DELAY <= MAXIMUM_BACKOFF_DELAY)
+					delay(BACKOFF_DELAY);
+				if (BACKOFF_DELAY == 2400000)
+					BACKOFF_DELAY += 1500000;
+				else
+					BACKOFF_DELAY = BACKOFF_DELAY * INCREMENT_BACKOFF_DELAY;
+			} else  throw new ServerException(URLErrorCodes.ERR_INVALID_UPLOAD_FILE_URL.name(), "Invalid Response Received From Google API for file Id : " + fileId + " | Error is : " + he.getContent());
+		} catch (Exception e) {
+			TelemetryManager.log("GoogleDriveUtil :: downloadFile :: Exception :: Error Occurred While Downloading Google Drive File having Id " + fileId + " : " + e.getMessage());
+			e.printStackTrace();
+			throw new ServerException(URLErrorCodes.ERR_INVALID_UPLOAD_FILE_URL.name(), "Invalid Response Received From Google API for file Id : " + fileId + " | Error is : " + e.getMessage());
+		}
+		return null;
+	}
+
+	public static void delay(int time) {
+		TelemetryManager.log("delay is called with : " + time);
+		try {
+			Thread.sleep(time);
+		} catch (Exception e) {
+
+		}
 	}
 }
