+++ platform-modules/url-manager/src/test/java/org/sunbird/url/util/GoogleDriveUtilTest.java
@@ -101,4 +101,13 @@ public class GoogleDriveUtilTest {
 		String driveUrl = "https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy1.pdf";
 		Long result = GoogleDriveUrlUtil.getSize(driveUrl);
 	}
+
+	@Test
+	public void testDownloadFileWithValidUrl() {
+		String downloadFolder = "/tmp/content/" + System.currentTimeMillis() + "_temp/do_123";
+		String driveUrl = "https://drive.google.com/uc?export=download&id=1nH-aHrU4cKcqFMI2HLHk8LtqxLctOwu9";
+		java.io.File appIconFile = GoogleDriveUrlUtil.downloadFile(driveUrl,downloadFolder);
+		assertTrue(appIconFile.exists());
+	}
+
 }
